requirements.txt
requests
langchain
langchain-community
langchain-openai
chromadb
openai
fastapi
uvicorn
selenium
beautifulsoup4
APScheduler
azure-storage-blob
pyshorteners
python-dotenv 
email-validator
ydata-profiling
pandasai
llama-index
pyyaml
numpy>=1.21,<1.25  
numba==0.57.0
python-multipart
scikit-learn
webdriver-manager
playwright
colorama
kneed
python-jose
pymongo
llama-index-embeddings-azure-openai
PyPDF2
pdfplumber
docx2txt
